refactor: extract output format modules from output/__init__.py
Splits 6 format converters into dedicated modules:
- json_fmt.py: to_json, export_json, framework summary builder
- sarif.py: to_sarif, export_sarif (SARIF 2.1.0)
- cyclonedx_fmt.py: to_cyclonedx, export_cyclonedx (CycloneDX 1.6)
- spdx_fmt.py: to_spdx, export_spdx (SPDX 3.0)
- badge.py: to_badge, export_badge, to_rsp_badge
- compliance_export.py: export_compliance_bundle

__init__.py: 2,784 → 1,662 lines (-40%). All symbols re-exported
for backward compatibility. 1,268 tests pass.
