+++ src/agent_bom/output/__init__.py
@@ -1198,789 +1198,6 @@ def _pct(part: int, total: int) -> str:
     return f"{round(part / total * 100)}%" if total > 0 else "—"
 
 
-def _build_remediation_json(report: AIBOMReport) -> list[dict]:
-    """Build JSON-serializable remediation plan with named assets and percentages."""
-    plan = build_remediation_plan(report.blast_radii)
-    total_agents = report.total_agents or 1
-
-    all_creds: set[str] = set()
-    all_tools: set[str] = set()
-    for br in report.blast_radii:
-        all_creds.update(br.exposed_credentials)
-        all_tools.update(t.name for t in br.exposed_tools)
-    total_creds = len(all_creds) or 1
-    total_tools = len(all_tools) or 1
-
-    result = []
-    for item in plan:
-        n_agents = len(item["agents"])
-        n_creds = len(item["creds"])
-        n_tools = len(item["tools"])
-        result.append(
-            {
-                "package": item["package"],
-                "ecosystem": item["ecosystem"],
-                "current_version": item["current"],
-                "fixed_version": item["fix"],
-                "severity": item["max_severity"].value,
-                "is_kev": item["has_kev"],
-                "impact_score": item["impact"],
-                "vulnerabilities": item["vulns"],
-                "affected_agents": item["agents"],
-                "agents_pct": round(n_agents / total_agents * 100),
-                "exposed_credentials": item["creds"],
-                "credentials_pct": round(n_creds / total_creds * 100) if n_creds else 0,
-                "reachable_tools": item["tools"],
-                "tools_pct": round(n_tools / total_tools * 100) if n_tools else 0,
-                "owasp_tags": item["owasp"],
-                "atlas_tags": item["atlas"],
-                "nist_ai_rmf_tags": item["nist"],
-                "owasp_mcp_tags": item["owasp_mcp"],
-                "owasp_agentic_tags": item["owasp_agentic"],
-                "eu_ai_act_tags": item["eu_ai_act"],
-                "nist_csf_tags": item["nist_csf"],
-                "iso_27001_tags": item["iso_27001"],
-                "soc2_tags": item["soc2"],
-                "cis_tags": item["cis"],
-                "references": item.get("references", []),
-                "risk_narrative": _risk_narrative(item),
-            }
-        )
-    return result
-
-
-def _risk_narrative(item: dict) -> str:
-    """Build plain-text risk narrative for a remediation item."""
-    vuln_id = item["vulns"][0] if item["vulns"] else "this vulnerability"
-    agents = ", ".join(item["agents"][:3]) or "affected agents"
-    creds = ", ".join(item["creds"][:3])
-    tools = ", ".join(item["tools"][:3])
-
-    parts = [f"If not remediated, an attacker exploiting {vuln_id}"]
-    if creds:
-        parts.append(f"can exfiltrate {creds}")
-    parts.append(f"via {agents}")
-    if tools:
-        parts.append(f"through {tools}")
-    return " ".join(parts) + "."
-
-
-# ─── JSON Output ────────────────────────────────────────────────────────────
-
-
-def _build_framework_summary(blast_radii: list[BlastRadius]) -> dict:
-    """Aggregate OWASP + ATLAS tag coverage across all blast radius findings."""
-    from collections import Counter
-
-    from agent_bom.atlas import ATLAS_TECHNIQUES
-    from agent_bom.cis_controls import CIS_CONTROLS
-    from agent_bom.eu_ai_act import EU_AI_ACT
-    from agent_bom.iso_27001 import ISO_27001
-    from agent_bom.nist_ai_rmf import NIST_AI_RMF
-    from agent_bom.nist_csf import NIST_CSF
-    from agent_bom.owasp import OWASP_LLM_TOP10
-    from agent_bom.owasp_agentic import OWASP_AGENTIC_TOP10
-    from agent_bom.owasp_mcp import OWASP_MCP_TOP10
-    from agent_bom.soc2 import SOC2_TSC
-
-    owasp_counts: Counter[str] = Counter()
-    atlas_counts: Counter[str] = Counter()
-    nist_counts: Counter[str] = Counter()
-    owasp_mcp_counts: Counter[str] = Counter()
-    owasp_agentic_counts: Counter[str] = Counter()
-    eu_ai_act_counts: Counter[str] = Counter()
-    nist_csf_counts: Counter[str] = Counter()
-    iso_27001_counts: Counter[str] = Counter()
-    soc2_counts: Counter[str] = Counter()
-    cis_counts: Counter[str] = Counter()
-    for br in blast_radii:
-        for tag in br.owasp_tags:
-            owasp_counts[tag] += 1
-        for tag in br.atlas_tags:
-            atlas_counts[tag] += 1
-        for tag in br.nist_ai_rmf_tags:
-            nist_counts[tag] += 1
-        for tag in br.owasp_mcp_tags:
-            owasp_mcp_counts[tag] += 1
-        for tag in br.owasp_agentic_tags:
-            owasp_agentic_counts[tag] += 1
-        for tag in br.eu_ai_act_tags:
-            eu_ai_act_counts[tag] += 1
-        for tag in br.nist_csf_tags:
-            nist_csf_counts[tag] += 1
-        for tag in br.iso_27001_tags:
-            iso_27001_counts[tag] += 1
-        for tag in br.soc2_tags:
-            soc2_counts[tag] += 1
-        for tag in br.cis_tags:
-            cis_counts[tag] += 1
-
-    return {
-        "owasp_llm_top10": [
-            {
-                "code": code,
-                "name": OWASP_LLM_TOP10[code],
-                "findings": owasp_counts.get(code, 0),
-                "triggered": code in owasp_counts,
-            }
-            for code in sorted(OWASP_LLM_TOP10.keys())
-        ],
-        "mitre_atlas": [
-            {
-                "technique_id": tid,
-                "name": ATLAS_TECHNIQUES[tid],
-                "findings": atlas_counts.get(tid, 0),
-                "triggered": tid in atlas_counts,
-            }
-            for tid in sorted(ATLAS_TECHNIQUES.keys())
-        ],
-        "nist_ai_rmf": [
-            {
-                "subcategory_id": sid,
-                "name": NIST_AI_RMF[sid],
-                "findings": nist_counts.get(sid, 0),
-                "triggered": sid in nist_counts,
-            }
-            for sid in sorted(NIST_AI_RMF.keys())
-        ],
-        "owasp_mcp_top10": [
-            {
-                "code": code,
-                "name": OWASP_MCP_TOP10[code],
-                "findings": owasp_mcp_counts.get(code, 0),
-                "triggered": code in owasp_mcp_counts,
-            }
-            for code in sorted(OWASP_MCP_TOP10.keys())
-        ],
-        "owasp_agentic_top10": [
-            {
-                "code": code,
-                "name": OWASP_AGENTIC_TOP10[code],
-                "findings": owasp_agentic_counts.get(code, 0),
-                "triggered": code in owasp_agentic_counts,
-            }
-            for code in sorted(OWASP_AGENTIC_TOP10.keys())
-        ],
-        "eu_ai_act": [
-            {
-                "code": code,
-                "name": EU_AI_ACT[code],
-                "findings": eu_ai_act_counts.get(code, 0),
-                "triggered": code in eu_ai_act_counts,
-            }
-            for code in sorted(EU_AI_ACT.keys())
-        ],
-        "nist_csf": [
-            {
-                "code": code,
-                "name": NIST_CSF[code],
-                "findings": nist_csf_counts.get(code, 0),
-                "triggered": code in nist_csf_counts,
-            }
-            for code in sorted(NIST_CSF.keys())
-        ],
-        "iso_27001": [
-            {
-                "code": code,
-                "name": ISO_27001[code],
-                "findings": iso_27001_counts.get(code, 0),
-                "triggered": code in iso_27001_counts,
-            }
-            for code in sorted(ISO_27001.keys())
-        ],
-        "soc2_tsc": [
-            {
-                "code": code,
-                "name": SOC2_TSC[code],
-                "findings": soc2_counts.get(code, 0),
-                "triggered": code in soc2_counts,
-            }
-            for code in sorted(SOC2_TSC.keys())
-        ],
-        "cis_controls": [
-            {
-                "code": code,
-                "name": CIS_CONTROLS[code],
-                "findings": cis_counts.get(code, 0),
-                "triggered": code in cis_counts,
-            }
-            for code in sorted(CIS_CONTROLS.keys())
-        ],
-        "total_owasp_triggered": sum(1 for c in owasp_counts if owasp_counts[c] > 0),
-        "total_atlas_triggered": sum(1 for c in atlas_counts if atlas_counts[c] > 0),
-        "total_nist_triggered": sum(1 for c in nist_counts if nist_counts[c] > 0),
-        "total_owasp_mcp_triggered": sum(1 for c in owasp_mcp_counts if owasp_mcp_counts[c] > 0),
-        "total_owasp_agentic_triggered": sum(1 for c in owasp_agentic_counts if owasp_agentic_counts[c] > 0),
-        "total_eu_ai_act_triggered": sum(1 for c in eu_ai_act_counts if eu_ai_act_counts[c] > 0),
-        "total_nist_csf_triggered": sum(1 for c in nist_csf_counts if nist_csf_counts[c] > 0),
-        "total_iso_27001_triggered": sum(1 for c in iso_27001_counts if iso_27001_counts[c] > 0),
-        "total_soc2_triggered": sum(1 for c in soc2_counts if soc2_counts[c] > 0),
-        "total_cis_triggered": sum(1 for c in cis_counts if cis_counts[c] > 0),
-    }
-
-
-def to_json(report: AIBOMReport) -> dict:
-    """Convert report to JSON-serializable dict."""
-    result = {
-        "document_type": "AI-BOM",
-        "spec_version": "1.0",
-        "ai_bom_version": report.tool_version,
-        "generated_at": report.generated_at.isoformat(),
-        "scan_sources": report.scan_sources,
-        "has_mcp_context": report.has_mcp_context,
-        "has_agent_context": report.has_agent_context,
-        "summary": {
-            "total_agents": report.total_agents,
-            "total_mcp_servers": report.total_servers,
-            "total_packages": report.total_packages,
-            "total_vulnerabilities": report.total_vulnerabilities,
-            "critical_findings": len(report.critical_vulns),
-        },
-        "agents": [
-            {
-                "name": agent.name,
-                "type": agent.agent_type.value,
-                "config_path": agent.config_path,
-                "source": agent.source,
-                "status": agent.status.value,
-                "mcp_servers": [
-                    {
-                        "name": server.name,
-                        "command": server.command,
-                        "args": server.args,
-                        "transport": server.transport.value,
-                        "url": server.url,
-                        "mcp_version": server.mcp_version,
-                        "has_credentials": server.has_credentials,
-                        "credential_env_vars": server.credential_names,
-                        "security_blocked": server.security_blocked,
-                        "security_warnings": server.security_warnings,
-                        "tools": [{"name": t.name, "description": t.description} for t in server.tools],
-                        "packages": [
-                            {
-                                "name": pkg.name,
-                                "version": pkg.version,
-                                "ecosystem": pkg.ecosystem,
-                                "purl": pkg.purl,
-                                "is_direct": pkg.is_direct,
-                                "parent_package": pkg.parent_package,
-                                "dependency_depth": pkg.dependency_depth,
-                                "resolved_from_registry": pkg.resolved_from_registry,
-                                "version_source": pkg.version_source,
-                                "registry_version": pkg.registry_version,
-                                "license": pkg.license,
-                                "license_expression": pkg.license_expression,
-                                "supplier": pkg.supplier,
-                                "author": pkg.author,
-                                "description": pkg.description,
-                                "homepage": pkg.homepage,
-                                "repository_url": pkg.repository_url,
-                                "download_url": pkg.download_url,
-                                "copyright_text": pkg.copyright_text,
-                                "deps_dev_resolved": pkg.deps_dev_resolved,
-                                "scorecard_score": pkg.scorecard_score,
-                                "scorecard_checks": pkg.scorecard_checks or None,
-                                "vulnerabilities": [
-                                    {
-                                        "id": v.id,
-                                        "summary": v.summary,
-                                        "severity": v.severity.value,
-                                        "cvss_score": v.cvss_score,
-                                        "epss_score": v.epss_score,
-                                        "epss_percentile": v.epss_percentile,
-                                        "is_kev": v.is_kev,
-                                        "kev_date_added": v.kev_date_added,
-                                        "cwe_ids": v.cwe_ids,
-                                        "fixed_version": v.fixed_version,
-                                        "references": v.references,
-                                        "nvd_published": v.nvd_published,
-                                        "nvd_modified": v.nvd_modified,
-                                        "nvd_status": v.nvd_status,
-                                        "vex_status": v.vex_status,
-                                        "vex_justification": v.vex_justification,
-                                        "compliance_tags": v.compliance_tags,
-                                    }
-                                    for v in pkg.vulnerabilities
-                                ],
-                            }
-                            for pkg in server.packages
-                        ],
-                        "permission_profile": (
-                            {
-                                "runs_as_root": server.permission_profile.runs_as_root,
-                                "container_privileged": server.permission_profile.container_privileged,
-                                "privilege_level": server.permission_profile.privilege_level,
-                                "tool_permissions": server.permission_profile.tool_permissions,
-                                "capabilities": server.permission_profile.capabilities,
-                                "network_access": server.permission_profile.network_access,
-                                "filesystem_write": server.permission_profile.filesystem_write,
-                                "shell_access": server.permission_profile.shell_access,
-                            }
-                            if server.permission_profile
-                            else None
-                        ),
-                    }
-                    for server in agent.mcp_servers
-                ],
-            }
-            for agent in report.agents
-        ],
-        "blast_radius": [
-            {
-                "risk_score": br.risk_score,
-                "vulnerability_id": br.vulnerability.id,
-                "severity": br.vulnerability.severity.value,
-                "cvss_score": br.vulnerability.cvss_score,
-                "epss_score": br.vulnerability.epss_score,
-                "is_kev": br.vulnerability.is_kev,
-                "nvd_status": br.vulnerability.nvd_status,
-                "compliance_tags": br.vulnerability.compliance_tags,
-                "package": f"{br.package.name}@{br.package.version}",
-                "ecosystem": br.package.ecosystem,
-                "is_malicious": br.package.is_malicious,
-                "malicious_reason": br.package.malicious_reason,
-                "scorecard_score": br.package.scorecard_score,
-                "affected_agents": [a.name for a in br.affected_agents],
-                "affected_servers": [s.name for s in br.affected_servers],
-                "exposed_credentials": br.exposed_credentials,
-                "exposed_tools": [t.name for t in br.exposed_tools],
-                "fixed_version": br.vulnerability.fixed_version,
-                "ai_risk_context": br.ai_risk_context,
-                "ai_summary": br.ai_summary,
-                "owasp_tags": br.owasp_tags,
-                "atlas_tags": br.atlas_tags,
-                "attack_tags": getattr(br, "attack_tags", []),
-                "nist_ai_rmf_tags": br.nist_ai_rmf_tags,
-                "owasp_mcp_tags": br.owasp_mcp_tags,
-                "owasp_agentic_tags": br.owasp_agentic_tags,
-                "eu_ai_act_tags": br.eu_ai_act_tags,
-                "nist_csf_tags": br.nist_csf_tags,
-                "iso_27001_tags": br.iso_27001_tags,
-                "soc2_tags": br.soc2_tags,
-                "cis_tags": br.cis_tags,
-            }
-            for br in report.blast_radii
-        ],
-        "threat_framework_summary": _build_framework_summary(report.blast_radii),
-        "remediation_plan": _build_remediation_json(report),
-    }
-
-    # AI enrichment fields (only when present)
-    if report.executive_summary:
-        result["executive_summary"] = report.executive_summary
-    if report.ai_threat_chains:
-        result["ai_threat_chains"] = report.ai_threat_chains
-    if report.mcp_config_analysis:
-        result["mcp_config_analysis"] = report.mcp_config_analysis
-
-    # Skill security audit (only when skill files were scanned)
-    if report.skill_audit_data:
-        result["skill_audit"] = report.skill_audit_data
-
-    # Trust assessment (only when skill files were scanned)
-    if report.trust_assessment_data:
-        result["trust_assessment"] = report.trust_assessment_data
-
-    if report.prompt_scan_data:
-        result["prompt_scan"] = report.prompt_scan_data
-
-    if report.model_files:
-        result["model_files"] = report.model_files
-
-    if report.model_provenance:
-        result["model_provenance"] = report.model_provenance
-
-    if report.enforcement_data:
-        result["enforcement"] = report.enforcement_data
-
-    if report.context_graph_data:
-        result["context_graph"] = report.context_graph_data
-
-    if report.license_report:
-        result["license_report"] = report.license_report
-
-    if report.vex_data:
-        result["vex"] = report.vex_data
-
-    if report.toxic_combinations:
-        result["toxic_combinations"] = report.toxic_combinations
-
-    if report.prioritized_findings:
-        result["prioritized_findings"] = report.prioritized_findings
-
-    if report.sast_data:
-        result["sast"] = report.sast_data
-
-    if report.cis_benchmark_data:
-        result["cis_benchmark"] = report.cis_benchmark_data
-
-    if report.snowflake_cis_benchmark_data:
-        result["snowflake_cis_benchmark"] = report.snowflake_cis_benchmark_data
-
-    if report.azure_cis_benchmark_data:
-        result["azure_cis_benchmark"] = report.azure_cis_benchmark_data
-
-    if report.gcp_cis_benchmark_data:
-        result["gcp_cis_benchmark"] = report.gcp_cis_benchmark_data
-
-    if report.databricks_cis_benchmark_data:
-        result["databricks_cis_benchmark"] = report.databricks_cis_benchmark_data
-
-    # Training pipeline lineage + dataset cards
-    if report.training_pipelines:
-        result["training_pipelines"] = report.training_pipelines
-    if report.dataset_cards:
-        result["dataset_cards"] = report.dataset_cards
-    if report.serving_configs:
-        result["serving_configs"] = report.serving_configs
-    if report.browser_extensions:
-        result["browser_extensions"] = report.browser_extensions
-
-    # Posture scorecard
-    from agent_bom.posture import (
-        compute_credential_risk_ranking,
-        compute_incident_correlation,
-        compute_posture_scorecard,
-    )
-
-    scorecard = compute_posture_scorecard(report)
-    result["posture_scorecard"] = scorecard.to_dict()
-    result["credential_risk_ranking"] = compute_credential_risk_ranking(report)
-    result["incident_correlation"] = compute_incident_correlation(report)
-
-    return result
-
-
-def export_json(report: AIBOMReport, output_path: str) -> None:
-    """Export report as JSON file."""
-    data = to_json(report)
-    Path(output_path).write_text(json.dumps(data, indent=2))
-
-
-# ─── CycloneDX Output ──────────────────────────────────────────────────────
-
-
-def _sanitize_bom_ref(raw: str) -> str:
-    """Sanitize a CycloneDX bom-ref to contain only valid characters.
-
-    CycloneDX 1.6 bom-ref should match ``^[a-zA-Z0-9._-]+$``.
-    Replace invalid characters (``@``, ``/``, spaces, etc.) with ``-``.
-    """
-    import re
-
-    return re.sub(r"[^a-zA-Z0-9._-]", "-", raw)
-
-
-def to_cyclonedx(report: AIBOMReport) -> dict:
-    """Build CycloneDX 1.6 dict from report."""
-    components = []
-    vulnerabilities_cdx = []
-    dependencies = []
-
-    comp_id = 0
-    bom_ref_map = {}
-
-    # Add agents as top-level components
-    for agent in report.agents:
-        agent_ref = _sanitize_bom_ref(f"agent-{agent.name}")
-        agent_deps = []
-
-        components.append(
-            {
-                "type": "application",
-                "bom-ref": agent_ref,
-                "name": agent.name,
-                "version": agent.version or "unknown",
-                "description": f"AI Agent ({agent.agent_type.value})",
-                "properties": [
-                    {"name": "agent-bom:type", "value": "ai-agent"},
-                    {"name": "agent-bom:config-path", "value": agent.config_path},
-                    {"name": "agent-bom:status", "value": agent.status.value},
-                ],
-            }
-        )
-
-        for server in agent.mcp_servers:
-            server_ref = _sanitize_bom_ref(f"mcp-server-{server.name}-{comp_id}")
-            comp_id += 1
-            server_deps = []
-
-            server_props = [
-                {"name": "agent-bom:type", "value": "mcp-server"},
-                {"name": "agent-bom:command", "value": server.command},
-                {"name": "agent-bom:transport", "value": server.transport.value},
-            ]
-            if server.has_credentials:
-                server_props.append({"name": "agent-bom:has-credentials", "value": "true"})
-
-            components.append(
-                {
-                    "type": "application",
-                    "bom-ref": server_ref,
-                    "name": server.name,
-                    "description": f"MCP Server ({server.transport.value})",
-                    "properties": server_props,
-                }
-            )
-            agent_deps.append(server_ref)
-
-            for pkg in server.packages:
-                pkg_ref = _sanitize_bom_ref(f"pkg-{pkg.ecosystem}-{pkg.name}-{pkg.version}-{comp_id}")
-                comp_id += 1
-
-                pkg_properties = [
-                    {"name": "agent-bom:ecosystem", "value": pkg.ecosystem},
-                    {"name": "agent-bom:is-direct", "value": str(pkg.is_direct).lower()},
-                    {"name": "agent-bom:dependency-depth", "value": str(pkg.dependency_depth)},
-                    {"name": "agent-bom:resolved-from-registry", "value": str(pkg.resolved_from_registry).lower()},
-                ]
-                if pkg.parent_package:
-                    pkg_properties.append({"name": "agent-bom:parent-package", "value": pkg.parent_package})
-                if pkg.scorecard_score is not None:
-                    pkg_properties.append({"name": "agent-bom:scorecard-score", "value": str(pkg.scorecard_score)})
-
-                pkg_component: dict = {
-                    "type": "library",
-                    "bom-ref": pkg_ref,
-                    "name": pkg.name,
-                    "version": pkg.version,
-                    "purl": pkg.purl,
-                    "properties": pkg_properties,
-                }
-                if pkg.license_expression or pkg.license:
-                    lic_id = pkg.license_expression or pkg.license
-                    pkg_component["licenses"] = [{"license": {"id": lic_id}}]
-                if pkg.supplier:
-                    pkg_component["supplier"] = {"name": pkg.supplier}
-                if pkg.author:
-                    pkg_component["author"] = pkg.author
-                if pkg.description:
-                    pkg_component["description"] = pkg.description
-                if pkg.copyright_text:
-                    pkg_component["copyright"] = pkg.copyright_text
-                # External references (homepage, repository, download)
-                ext_refs = []
-                if pkg.homepage:
-                    ext_refs.append({"type": "website", "url": pkg.homepage})
-                if pkg.repository_url:
-                    ext_refs.append({"type": "vcs", "url": pkg.repository_url})
-                if pkg.download_url:
-                    ext_refs.append({"type": "distribution", "url": pkg.download_url})
-                if ext_refs:
-                    pkg_component["externalReferences"] = ext_refs
-                components.append(pkg_component)
-                server_deps.append(pkg_ref)
-                bom_ref_map[f"{pkg.ecosystem}:{pkg.name}@{pkg.version}"] = pkg_ref
-
-                # Add vulnerabilities
-                for vuln in pkg.vulnerabilities:
-                    ratings: list[dict[str, object]] = []
-                    if vuln.cvss_score:
-                        ratings.append(
-                            {
-                                "score": vuln.cvss_score,
-                                "severity": vuln.severity.value,
-                                "method": "CVSSv3",
-                            }
-                        )
-                    else:
-                        ratings.append(
-                            {
-                                "severity": vuln.severity.value,
-                            }
-                        )
-                    vuln_entry: dict[str, object] = {
-                        "id": vuln.id,
-                        "description": vuln.summary,
-                        "source": {"name": "OSV", "url": f"https://osv.dev/vulnerability/{vuln.id}"},
-                        "ratings": ratings,
-                        "affects": [{"ref": pkg_ref}],
-                    }
-                    if vuln.fixed_version:
-                        vuln_entry["recommendation"] = f"Upgrade to {vuln.fixed_version}"
-                    if vuln.vex_status:
-                        # CycloneDX VEX analysis mapping
-                        _cdx_state_map = {
-                            "affected": "exploitable",
-                            "not_affected": "not_affected",
-                            "fixed": "resolved",
-                            "under_investigation": "in_triage",
-                        }
-                        analysis_dict: dict[str, str] = {
-                            "state": _cdx_state_map.get(vuln.vex_status, "in_triage"),
-                        }
-                        if vuln.vex_justification:
-                            analysis_dict["justification"] = vuln.vex_justification
-                        vuln_entry["analysis"] = analysis_dict
-                    vulnerabilities_cdx.append(vuln_entry)
-
-            dependencies.append({"ref": server_ref, "dependsOn": server_deps})
-        dependencies.append({"ref": agent_ref, "dependsOn": agent_deps})
-
-    cdx = {
-        "bomFormat": "CycloneDX",
-        "specVersion": "1.6",
-        "serialNumber": f"urn:uuid:{uuid4()}",
-        "version": 1,
-        "metadata": {
-            "timestamp": report.generated_at.isoformat(),
-            "tools": {
-                "components": [
-                    {
-                        "type": "application",
-                        "name": "agent-bom",
-                        "version": report.tool_version,
-                        "description": "AI Bill of Materials generator for AI agents and MCP servers",
-                    }
-                ]
-            },
-            "properties": [
-                {"name": "agent-bom:total-agents", "value": str(report.total_agents)},
-                {"name": "agent-bom:total-mcp-servers", "value": str(report.total_servers)},
-                {"name": "agent-bom:total-vulnerabilities", "value": str(report.total_vulnerabilities)},
-            ],
-        },
-        "components": components,
-        "dependencies": dependencies,
-    }
-
-    if vulnerabilities_cdx:
-        cdx["vulnerabilities"] = vulnerabilities_cdx
-
-    return cdx
-
-
-def export_cyclonedx(report: AIBOMReport, output_path: str) -> None:
-    """Export report as CycloneDX 1.6 JSON file."""
-    cdx = to_cyclonedx(report)
-    Path(output_path).write_text(json.dumps(cdx, indent=2))
-
-
-# ─── SARIF Output ──────────────────────────────────────────────────────────
-
-_SARIF_SEVERITY_MAP = {
-    Severity.CRITICAL: "error",
-    Severity.HIGH: "error",
-    Severity.MEDIUM: "warning",
-    Severity.LOW: "note",
-    Severity.NONE: "none",
-}
-
-
-def to_sarif(report: AIBOMReport) -> dict:
-    """Convert report to SARIF 2.1.0 dict for GitHub Security tab."""
-    rules = []
-    results = []
-    seen_rule_ids: set[str] = set()
-
-    for br in report.blast_radii:
-        vuln = br.vulnerability
-        rule_id = vuln.id
-        level = _SARIF_SEVERITY_MAP.get(vuln.severity, "warning")
-
-        if rule_id not in seen_rule_ids:
-            seen_rule_ids.add(rule_id)
-            rule: dict = {
-                "id": rule_id,
-                "shortDescription": {"text": f"{vuln.severity.value.upper()}: {vuln.id} in {br.package.name}@{br.package.version}"},
-                "fullDescription": {"text": vuln.summary or f"Vulnerability {vuln.id}"},
-                "helpUri": f"https://osv.dev/vulnerability/{vuln.id}",
-                "defaultConfiguration": {"level": level},
-            }
-            if vuln.cwe_ids:
-                rule["properties"] = {"tags": vuln.cwe_ids}
-            rules.append(rule)
-
-        affected = ", ".join(a.name for a in br.affected_agents)
-        message_text = f"{vuln.id} ({vuln.severity.value}) in {br.package.name}@{br.package.version}. Affects agents: {affected}."
-        if vuln.fixed_version:
-            message_text += f" Fix: upgrade to {vuln.fixed_version}."
-
-        config_path = br.affected_agents[0].config_path if br.affected_agents else "unknown"
-
-        # SARIF 2.1.0: fingerprints for dedup, region for location, kind for classification
-        import hashlib
-
-        fp_input = f"{rule_id}:{br.package.name}:{br.package.version}:{config_path}"
-        fingerprint = hashlib.sha256(fp_input.encode()).hexdigest()
-
-        result: dict = {
-            "ruleId": rule_id,
-            "level": level,
-            "kind": "fail",
-            "message": {"text": message_text},
-            "fingerprints": {
-                "agent-bom/v1": fingerprint,
-            },
-            "locations": [
-                {
-                    "physicalLocation": {
-                        "artifactLocation": {
-                            "uri": config_path,
-                            "uriBaseId": "%SRCROOT%",
-                        },
-                        "region": {
-                            "startLine": 1,
-                            "startColumn": 1,
-                        },
-                    },
-                }
-            ],
-        }
-        if (
-            br.owasp_tags
-            or br.atlas_tags
-            or br.nist_ai_rmf_tags
-            or br.owasp_mcp_tags
-            or br.owasp_agentic_tags
-            or br.eu_ai_act_tags
-            or br.nist_csf_tags
-            or br.iso_27001_tags
-            or br.soc2_tags
-            or br.cis_tags
-        ):
-            result["properties"] = {
-                "owasp_tags": br.owasp_tags,
-                "atlas_tags": br.atlas_tags,
-                "attack_tags": getattr(br, "attack_tags", []),
-                "nist_ai_rmf_tags": br.nist_ai_rmf_tags,
-                "owasp_mcp_tags": br.owasp_mcp_tags,
-                "owasp_agentic_tags": br.owasp_agentic_tags,
-                "eu_ai_act_tags": br.eu_ai_act_tags,
-                "nist_csf_tags": br.nist_csf_tags,
-                "iso_27001_tags": br.iso_27001_tags,
-                "soc2_tags": br.soc2_tags,
-                "cis_tags": br.cis_tags,
-                "blast_score": br.risk_score,
-                "exposed_credentials": br.exposed_credentials,
-            }
-        results.append(result)
-
-    return {
-        "$schema": "https://raw.githubusercontent.com/oasis-tcs/sarif-spec/main/sarif-2.1/schema/sarif-schema-2.1.0.json",
-        "version": "2.1.0",
-        "runs": [
-            {
-                "tool": {
-                    "driver": {
-                        "name": "agent-bom",
-                        "version": report.tool_version,
-                        "informationUri": "https://github.com/msaad00/agent-bom",
-                        "rules": rules,
-                    }
-                },
-                "results": results,
-            }
-        ],
-    }
-
-
-def export_sarif(report: AIBOMReport, output_path: str) -> None:
-    """Export report as SARIF 2.1.0 JSON file."""
-    data = to_sarif(report)
-    Path(output_path).write_text(json.dumps(data, indent=2))
-
-
 # ─── Compact Output (default mode) ───────────────────────────────────────────
 
 
@@ -2357,194 +1574,38 @@ def print_severity_chart(report: AIBOMReport) -> None:
     console.print()
 
 
-# ─── SPDX 3.0 Output ────────────────────────────────────────────────────────
+# ─── Delegated format modules ────────────────────────────────────────────────
+# Each format lives in its own module; re-exported here for backward compat.
 
+from agent_bom.output.badge import (  # noqa: E402
+    export_badge,  # noqa: F401
+    to_badge,  # noqa: F401
+    to_rsp_badge,  # noqa: F401
+)
+from agent_bom.output.compliance_export import (  # noqa: E402
+    export_compliance_bundle,  # noqa: F401
+)
+from agent_bom.output.cyclonedx_fmt import (  # noqa: E402
+    export_cyclonedx,  # noqa: F401
+    to_cyclonedx,  # noqa: F401
+)
+from agent_bom.output.json_fmt import (  # noqa: E402
+    _build_framework_summary,  # noqa: F401
+    _build_remediation_json,  # noqa: F401
+    _risk_narrative,  # noqa: F401
+    export_json,  # noqa: F401
+    to_json,  # noqa: F401
+)
+from agent_bom.output.sarif import (  # noqa: E402
+    export_sarif,  # noqa: F401
+    to_sarif,  # noqa: F401
+)
+from agent_bom.output.spdx_fmt import (  # noqa: E402
+    export_spdx,  # noqa: F401
+    to_spdx,  # noqa: F401
+)
 
-def to_spdx(report: AIBOMReport) -> dict:
-    """Build an SPDX 3.0 (JSON-LD) dict from report.
-
-    Follows the SPDX 3.0 AI BOM profile where applicable:
-    - Each agent becomes an /AI element
-    - Each package becomes a /Package element
-    - Vulnerabilities become /security/VulnAssessmentRelationship elements
-    - Dependency edges become DEPENDS_ON relationships
-    """
-    from datetime import timezone
-
-    spdx_id_counter = [0]
-
-    def _next_id(prefix: str = "SPDXRef") -> str:
-        spdx_id_counter[0] += 1
-        return f"{prefix}-{spdx_id_counter[0]}"
-
-    elements: list[dict[str, Any]] = []
-    relationships: list[dict[str, Any]] = []
-    document_id = _next_id("SPDXRef-DOCUMENT")
-
-    # Document / CreationInfo
-    creation_info = {
-        "specVersion": "3.0.0",
-        "created": report.generated_at.astimezone(timezone.utc).strftime("%Y-%m-%dT%H:%M:%SZ")
-        if report.generated_at.tzinfo
-        else report.generated_at.strftime("%Y-%m-%dT%H:%M:%SZ") + "Z",
-        "createdBy": [
-            {
-                "type": "Tool",
-                "name": f"agent-bom {report.tool_version}",
-                "externalIdentifier": [
-                    {
-                        "type": "PackageURL",
-                        "identifier": f"pkg:pypi/agent-bom@{report.tool_version}",
-                    }
-                ],
-            }
-        ],
-    }
-
-    pkg_ref_map: dict[str, str] = {}  # ecosystem:name@version → spdxId
-
-    for agent in report.agents:
-        agent_id = _next_id("SPDXRef-Agent")
-
-        agent_element: dict[str, Any] = {
-            "type": "SOFTWARE_PACKAGE",
-            "spdxId": agent_id,
-            "name": agent.name,
-            "primaryPurpose": "APPLICATION",
-            "description": f"AI Agent ({agent.agent_type.value})",
-            "annotation": [
-                {
-                    "type": "Annotation",
-                    "annotationType": "OTHER",
-                    "subject": agent_id,
-                    "statement": f"agent-bom:ai-agent-type={agent.agent_type.value}",
-                }
-            ],
-        }
-        if agent.config_path:
-            agent_element["comment"] = f"config_path: {agent.config_path}, status: {agent.status.value}"
-        if agent.source:
-            agent_element["originatedBy"] = agent.source
-        elements.append(agent_element)
-
-        for server in agent.mcp_servers:
-            server_id = _next_id("SPDXRef-MCPServer")
-
-            server_element = {
-                "type": "SOFTWARE_PACKAGE",
-                "spdxId": server_id,
-                "name": server.name,
-                "primaryPurpose": "APPLICATION",
-                "description": f"MCP Server ({server.transport.value})",
-            }
-            if server.mcp_version:
-                server_element["versionInfo"] = server.mcp_version
-            elements.append(server_element)
-
-            relationships.append(
-                {
-                    "type": "Relationship",
-                    "spdxId": _next_id("SPDXRef-Rel"),
-                    "relationshipType": "CONTAINS",
-                    "from": agent_id,
-                    "to": [server_id],
-                }
-            )
-
-            for pkg in server.packages:
-                pkg_key = f"{pkg.ecosystem}:{pkg.name}@{pkg.version}"
-                if pkg_key not in pkg_ref_map:
-                    pkg_id = _next_id("SPDXRef-Pkg")
-                    pkg_ref_map[pkg_key] = pkg_id
-
-                    pkg_element: dict[str, object] = {
-                        "type": "SOFTWARE_PACKAGE",
-                        "spdxId": pkg_id,
-                        "name": pkg.name,
-                        "versionInfo": pkg.version,
-                        "primaryPurpose": "LIBRARY",
-                    }
-                    if pkg.purl:
-                        pkg_element["externalIdentifier"] = [{"type": "PackageURL", "identifier": pkg.purl}]
-                    if pkg.license_expression or pkg.license:
-                        pkg_element["declaredLicense"] = pkg.license_expression or pkg.license
-                    if pkg.supplier:
-                        pkg_element["supplier"] = pkg.supplier
-                    if pkg.description:
-                        pkg_element["description"] = pkg.description[:300]
-                    if pkg.homepage:
-                        pkg_element["homepage"] = pkg.homepage
-                    if pkg.download_url:
-                        pkg_element["downloadLocation"] = pkg.download_url
-                    if pkg.copyright_text:
-                        pkg_element["copyrightText"] = pkg.copyright_text
-                    elements.append(pkg_element)
-
-                pkg_id = pkg_ref_map[pkg_key]
-                relationships.append(
-                    {
-                        "type": "Relationship",
-                        "spdxId": _next_id("SPDXRef-Rel"),
-                        "relationshipType": "DEPENDS_ON",
-                        "from": server_id,
-                        "to": [pkg_id],
-                    }
-                )
-
-                # Security relationships for each vulnerability
-                for vuln in pkg.vulnerabilities:
-                    vuln_element_id = _next_id("SPDXRef-Vuln")
-                    vuln_element: dict[str, object] = {
-                        "type": "security/Vulnerability",
-                        "spdxId": vuln_element_id,
-                        "name": vuln.id,
-                        "description": vuln.summary or "",
-                        "externalIdentifier": [{"type": "cve", "identifier": vuln.id}] if vuln.id.startswith("CVE-") else [],
-                    }
-                    if vuln.cvss_score is not None:
-                        vuln_element["assessedElement"] = pkg_id
-                        vuln_element["score"] = {
-                            "method": "CVSS_3",
-                            "score": vuln.cvss_score,
-                            "severity": vuln.severity.value,
-                        }
-                    elements.append(vuln_element)
-
-                    assessment_id = _next_id("SPDXRef-VulnAssessment")
-                    assessment = {
-                        "type": "security/VulnAssessmentRelationship",
-                        "spdxId": assessment_id,
-                        "relationshipType": "AFFECTS",
-                        "from": vuln_element_id,
-                        "to": [pkg_id],
-                        "severity": vuln.severity.value,
-                    }
-                    if vuln.fixed_version:
-                        assessment["remediation"] = f"Upgrade to {vuln.fixed_version}"
-                    if vuln.is_kev:
-                        assessment["comment"] = "CISA KEV: actively exploited in the wild"
-                    relationships.append(assessment)
-
-    return {
-        "spdxVersion": "SPDX-3.0",
-        "dataLicense": "CC0-1.0",
-        "SPDXID": document_id,
-        "name": f"agent-bom-{report.generated_at.strftime('%Y%m%d-%H%M%S')}",
-        "creationInfo": creation_info,
-        "elements": elements,
-        "relationships": relationships,
-        "comment": (
-            f"AI Bill of Materials generated by agent-bom {report.tool_version}. "
-            f"Covers {report.total_agents} agent(s), {report.total_servers} MCP server(s), "
-            f"{report.total_packages} package(s), {report.total_vulnerabilities} vulnerability/ies."
-        ),
-    }
-
-
-def export_spdx(report: AIBOMReport, output_path: str) -> None:
-    """Export report as SPDX 3.0 JSON-LD file."""
-    data = to_spdx(report)
-    Path(output_path).write_text(json.dumps(data, indent=2))
+# ─── HTML Output (delegated to html.py) ──────────────────────────────────────
 
 
 def to_html(report: AIBOMReport, blast_radii: list | None = None) -> str:
@@ -2561,7 +1622,7 @@ def export_html(report: AIBOMReport, output_path: str, blast_radii: list | None 
     _export_html(report, output_path, blast_radii or [])
 
 
-# ─── Prometheus Output ───────────────────────────────────────────────────────
+# ─── Prometheus Output (delegated to prometheus.py) ──────────────────────────
 
 
 def to_prometheus(report: AIBOMReport, blast_radii: list | None = None) -> str:
@@ -2600,185 +1661,3 @@ def push_otlp(
     from agent_bom.output.prometheus import push_otlp as _push_otlp
 
     _push_otlp(endpoint, report, blast_radii or [])
-
-
-# ─── Badge Output (shields.io endpoint format) ───────────────────────────
-
-
-def to_badge(report: AIBOMReport) -> dict:
-    """Generate a shields.io endpoint badge JSON from scan results.
-
-    The output follows the shields.io endpoint schema:
-    https://shields.io/badges/endpoint-badge
-
-    Use with: https://img.shields.io/endpoint?url=<badge-json-url>
-    """
-    # Collect all vulnerabilities from agents → servers → packages
-    all_vulns = []
-    for agent in report.agents:
-        for server in agent.mcp_servers:
-            for pkg in server.packages:
-                all_vulns.extend(pkg.vulnerabilities)
-    critical = len([v for v in all_vulns if v.severity == Severity.CRITICAL])
-    high = len([v for v in all_vulns if v.severity == Severity.HIGH])
-    total = len(all_vulns)
-
-    if critical > 0:
-        color = "critical"
-        message = f"{critical} critical, {high} high"
-    elif high > 0:
-        color = "orange"
-        message = f"{high} high, {total} total"
-    elif total > 0:
-        color = "yellow"
-        message = f"{total} findings"
-    else:
-        color = "brightgreen"
-        message = "clean"
-
-    return {
-        "schemaVersion": 1,
-        "label": "agent-bom",
-        "message": message,
-        "color": color,
-        "namedLogo": "shield",
-    }
-
-
-def export_badge(report: AIBOMReport, output_path: str) -> None:
-    """Export shields.io endpoint badge JSON to file."""
-    data = to_badge(report)
-    Path(output_path).write_text(json.dumps(data, indent=2))
-
-
-def to_rsp_badge(report: AIBOMReport) -> dict:
-    """Generate Anthropic RSP v3.0 alignment badge (shields.io JSON endpoint format).
-
-    Checks whether any Claude agents are present and whether they have
-    critical/high vulnerabilities in their supply chain.
-    """
-    from agent_bom.models import AgentType
-
-    claude_types = {AgentType.CLAUDE_DESKTOP, AgentType.CLAUDE_CODE}
-    claude_agents = [a for a in report.agents if a.agent_type in claude_types]
-
-    if not claude_agents:
-        return {
-            "schemaVersion": 1,
-            "label": "Anthropic RSP",
-            "message": "RSP n/a",
-            "color": "lightgrey",
-        }
-
-    has_vulns = any(br.vulnerability.severity in (Severity.CRITICAL, Severity.HIGH) for br in report.blast_radii)
-
-    if has_vulns:
-        return {
-            "schemaVersion": 1,
-            "label": "Anthropic RSP",
-            "message": "RSP review needed",
-            "color": "orange",
-        }
-
-    return {
-        "schemaVersion": 1,
-        "label": "Anthropic RSP",
-        "message": "RSP v3.0 aligned",
-        "color": "brightgreen",
-    }
-
-
-def export_compliance_bundle(
-    report: AIBOMReport,
-    framework: str,
-    output_path: str,
-) -> str:
-    """Export CMMC/FedRAMP/NIST-AI-RMF compliance evidence bundle as ZIP.
-
-    Returns the path to the written ZIP file.
-    """
-    import zipfile
-
-    cmmc_control_map = {
-        "CM-8": "Configuration Management — Component Inventory",
-        "SI-2": "System & Information Integrity — Flaw Remediation",
-        "SR-3": "Supply Chain Risk Management — Supply Chain Controls",
-        "RA-3": "Risk Assessment — Risk Assessment",
-        "AU-2": "Audit & Accountability — Event Logging",
-    }
-
-    # Build SBOM (CycloneDX)
-    sbom_data = to_cyclonedx(report)
-
-    # Build vulnerability report
-    vuln_entries = []
-    for br in report.blast_radii:
-        vuln_entries.append(
-            {
-                "id": br.vulnerability.id,
-                "severity": br.vulnerability.severity.value,
-                "package": br.package.name,
-                "version": br.package.version,
-                "fixed_version": br.vulnerability.fixed_version,
-                "risk_score": br.risk_score,
-                "affected_agents": [a.name for a in br.affected_agents],
-                "affected_servers": [s.name for s in br.affected_servers],
-            }
-        )
-
-    # Build policy results
-    policy_results = {
-        "framework": framework,
-        "scan_date": report.generated_at.isoformat(),
-        "tool_version": report.tool_version,
-        "total_agents": report.total_agents,
-        "total_servers": report.total_servers,
-        "total_packages": report.total_packages,
-        "total_vulnerabilities": report.total_vulnerabilities,
-        "critical_count": len(report.critical_vulns),
-    }
-
-    # Build control mapping
-    control_mapping = {}
-    for ctrl_id, ctrl_desc in cmmc_control_map.items():
-        findings = []
-        if ctrl_id == "CM-8":
-            findings = [{"type": "inventory", "count": report.total_packages}]
-        elif ctrl_id == "SI-2":
-            findings = [{"type": "vulnerabilities", "count": report.total_vulnerabilities}]
-        elif ctrl_id == "SR-3":
-            findings = [{"type": "supply_chain", "servers": report.total_servers}]
-        elif ctrl_id == "RA-3":
-            findings = [{"type": "risk_assessment", "blast_radii": len(report.blast_radii)}]
-        elif ctrl_id == "AU-2":
-            findings = [{"type": "audit_trail", "scan_date": report.generated_at.isoformat()}]
-        control_mapping[ctrl_id] = {
-            "description": ctrl_desc,
-            "status": "pass" if not findings or report.total_vulnerabilities == 0 else "review",
-            "findings": findings,
-        }
-
-    # Executive summary text
-    summary_lines = [
-        f"Compliance Evidence Bundle — {framework.upper()}",
-        f"Generated: {report.generated_at.isoformat()}",
-        f"Tool: agent-bom v{report.tool_version}",
-        "",
-        f"Agents scanned: {report.total_agents}",
-        f"MCP servers: {report.total_servers}",
-        f"Packages inventoried: {report.total_packages}",
-        f"Vulnerabilities found: {report.total_vulnerabilities}",
-        f"Critical findings: {len(report.critical_vulns)}",
-        "",
-        "Controls mapped: " + ", ".join(cmmc_control_map.keys()),
-    ]
-
-    zip_path = output_path if output_path.endswith(".zip") else output_path + ".zip"
-    with zipfile.ZipFile(zip_path, "w", zipfile.ZIP_DEFLATED) as zf:
-        zf.writestr("sbom.cdx.json", json.dumps(sbom_data, indent=2, default=str))
-        zf.writestr("vulnerability_report.json", json.dumps(vuln_entries, indent=2, default=str))
-        zf.writestr("policy_results.json", json.dumps(policy_results, indent=2, default=str))
-        zf.writestr("compliance_mapping.json", json.dumps(control_mapping, indent=2, default=str))
-        zf.writestr("summary.txt", "\n".join(summary_lines))
-
-    return zip_path
